+++ src/tranker.c
@@ -2,7 +2,7 @@
 
 #include <stdio.h>
 
-void add_event(event_t *event_list, int *event_counter, event_t *event) {
+void add_event(event_t *event_list, uint8_t *event_counter, event_t *event) {
 	event_list[*event_counter] = *event;
 	event_counter++;
 }
@@ -12,7 +12,7 @@ void *print_fuck(void *data) {
 	return NULL;
 }
 
-void run_events(event_t *event_list, int event_counter) {
+void run_events(event_t *event_list, uint8_t event_counter) {
 	int i;
 	for (i = 0; i < event_counter; i++) {
 		event_list[i].event_function(event_list[i].data);
